requirements/requirements.txt
python-dotenv==0.21.1
asgiref==3.6.0
Django==3.2.15
pytz==2022.7.1
sqlparse==0.4.3
pillow==9.4.0
django-cleanup==7.0.0
sorl-thumbnail==12.9.0